pyproject.toml
[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'

[tool.poetry]
name = "hijax"
version = "0.1.0"
description = "An experiment framework for Haiku and Jax"
authors = ["Angus Turner <angusturner27@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
dm-haiku = "^0.0.5"
optax = "^0.0.9"
wandb = "^0.12.6"
hydra-core = "^1.1.1"
mypy = "^0.910"
black = "^21.10b0"
tqdm = "^4.62.3"
torch = "^1.10.0"
numpy = "^1.21.4"
jax = "^0.2.24"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
